Renamed the discovery package to internal in the wemo binding. See openhab/static-code-analysis#100
