+++ extensions/binding/org.eclipse.smarthome.binding.wemo/OSGI-INF/WemoDiscoveryParticipant.xml
@@ -9,7 +9,7 @@
 
 -->
 <scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true" name="org.eclipse.smarthome.binding.wemo.discovery.WemoDiscoveryParticipant">
-   <implementation class="org.eclipse.smarthome.binding.wemo.discovery.WemoDiscoveryParticipant"/>
+   <implementation class="org.eclipse.smarthome.binding.wemo.internal.discovery.WemoDiscoveryParticipant"/>
    <service>
       <provide interface="org.eclipse.smarthome.config.discovery.UpnpDiscoveryParticipant"/>
    </service>
