+++ extensions/binding/org.eclipse.smarthome.binding.wemo/OSGI-INF/WemoDiscoveryService.xml
@@ -9,7 +9,7 @@
 
 -->
 <scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true" name="org.eclipse.smarthome.binding.wemo.discovery.WemoDiscoveryService">
-   <implementation class="org.eclipse.smarthome.binding.wemo.discovery.WemoDiscoveryService"/>
+   <implementation class="org.eclipse.smarthome.binding.wemo.internal.discovery.WemoDiscoveryService"/>
    <reference bind="setUpnpService" cardinality="1..1" interface="org.jupnp.UpnpService" name="UpnpService" policy="static" unbind="unsetUpnpService"/>
    <service>
       <provide interface="org.eclipse.smarthome.config.discovery.DiscoveryService"/>
