+++ extensions/binding/org.eclipse.smarthome.binding.wemo/src/main/java/org/eclipse/smarthome/binding/wemo/internal/discovery/WemoDiscoveryParticipant.java
@@ -5,7 +5,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.eclipse.smarthome.binding.wemo.discovery;
+package org.eclipse.smarthome.binding.wemo.internal.discovery;
 
 import static org.eclipse.smarthome.binding.wemo.WemoBindingConstants.*;
